Migrate to custom hook
